Deduplicate C++ run.py code.
Also shifts responsibility for producing JSON timings to the C++ tool
implementations. In practice, this is done by the shared main.h file.
This can be used to provide measurements of sub-components of the
evaluation, such as the time taken by the 'prepare' method.
